requirements.txt
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.11.4
django-debug-toolbar==1.8
django-import-export==0.5.1
django-nested-inline==0.3.7
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
odfpy==1.3.5
openpyxl==2.4.8
psycopg2==2.7.3
python-decouple==3.1
pytz==2017.2
PyYAML==3.12
sqlparse==0.2.3
tablib==0.11.5
unicodecsv==0.14.1
whitenoise==3.3.0
xlrd==1.0.0
xlwt==1.2.0
